Revert "need to update the movie model amd expose the api"
