--- Backend/server/src/main/java/com/backend/server/Model/DataTheater.java
@@ -1,24 +1,0 @@
-package com.backend.server.Model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Document(collection = "movies")
-public class DataTheater.java {
-    @Id
-    private int movieId;
-    private String title;
-    private String duration;
-    private String year;
-    private String posterUrl;
-    private String theater;
-    private String status;
-}
