+++ Backend/server/src/main/java/com/backend/server/Model/Movies.java
@@ -11,7 +11,7 @@ import org.springframework.data.mongodb.core.mapping.Document;
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-@Document(collection = "movies") //creates the collection in the mongodb 
+@Document(collection = "movies")
 public class Movies {
     @Id
     private int movieId;
