--- Backend/server/src/main/java/com/backend/server/Model/ShowTime.java
@@ -1,21 +1,0 @@
-package com.backend.server.Model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Document(collection = "showtimes")
-public class Movies {
-    @Id
-    private int showtime_id;
-    private String start_time;
-    private int theater_id;
-    private String end_time;
-}
