+++ Backend/server/src/main/java/com/backend/server/controller/MovieApis.java
@@ -34,8 +34,4 @@ public class MovieApis {
     public Movies addMovie(@RequestBody Movies movie){
         return movieService.addMovie(movie);
     }
-    @PostMapping("/updateMovie")
-    public Movies updateMovie(@RequestBody Movies movie){
-        return movieService.updateMovie(movie);
-    }
 }
