--- Backend/server/src/main/java/com/backend/server/repository/TicketTableRepository.java
@@ -1,12 +1,0 @@
-package com.backend.server.repository;
-
-import com.backend.server.Model.Movies;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface TicketTableRepository extends MongoRepository<Movies,Integer> {
-    List<Movies> findByStatus(String status);
-}
