--- Backend/server/src/main/java/com/backend/server/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.backend.server.repository;
-
-import com.backend.server.Model.User;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends MongoRepository<User,String> {
-
-
-}
