--- Backend/server/src/main/java/com/backend/server/service/CurrentUserDetails.java
@@ -1,26 +1,0 @@
-package com.backend.server.service;
-
-import com.backend.server.Model.User;
-import com.backend.server.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import java.util.Optional;
-
-@Service
-public class CurrentUserDetails implements UserDetailsService {
-    @Autowired
-    UserRepository userRepository;
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-
-        Optional<User> user=userRepository.findById(username);
-        if(user==null){
-            throw new UsernameNotFoundException("User not found");
-        }
-        return new CurrentUser(user);
-    }
-}
