+++ Backend/server/src/main/java/com/backend/server/service/MovieService.java
@@ -11,6 +11,4 @@ public interface MovieService {
     Movies addMovie(Movies movie);
 
     List<Movies> findByStatus(String status);
-    Movies updateMovie(Movies movie);
-    
 }
