+++ Backend/server/src/main/java/com/backend/server/service/MovieServiceImpl.java
@@ -32,12 +32,6 @@ public class MovieServiceImpl implements MovieService{
     }
 
     @Override
-    public Movies updateMovie(Movies movie) {
-        Movies mv=movieRepository.save(movie);
-        return mv;
-    }
-
-    @Override
     public List<Movies> findByStatus(String status) {
         List<Movies> movies=movieRepository.findByStatus(status);
         System.out.println(movies.size());
