--- Backend/server/src/main/resources/application.properties
@@ -1,18 +1,0 @@
-
-# MongoDB Configuration
-spring.data.mongodb.database=bookMyShow
-#spring.data.mongodb.authentication-database=admin
-#spring.data.mongodb.username=admin
-#spring.data.mongodb.password=password
-spring.data.mongodb.port=27017
-spring.data.mongodb.host=localhost
-
-
-#spring.security.user.name=user
-#spring.security.user.password=user
-#spring.security.user.roles=manager
-
-
-
-# Swagger Configuration
-spring.mvc.pathmatch.matching-strategy = ANT_PATH_MATCHER
