--- Frontend/src/Components/AgeRangeSelector
@@ -1,25 +1,0 @@
-import React, { useState } from 'react';
-
-function AgeRangeSelector() {
-  const [selectedAgeRange, setSelectedAgeRange] = useState('');
-
-  const handleAgeRangeChange = (event) => {
-    setSelectedAgeRange(event.target.value);
-  };
-
-  return (
-    <div>
-      <h2>Select Age Range</h2>
-      <select value={selectedAgeRange} onChange={handleAgeRangeChange}>
-        <option value="">Select Age Range</option>
-        <option value="0-18">0-18</option>
-        <option value="19-30">19-30</option>
-        <option value="31-50">31-50</option>
-        <option value="51-100">51-100</option>
-      </select>
-      {selectedAgeRange && <p>You selected: {selectedAgeRange}</p>}
-    </div>
-  );
-}
-
-export default AgeRangeSelector;
