+++ Frontend/src/Components/AppBar.js
@@ -7,15 +7,11 @@ export default function AppBar({dataFromAppBar}){
     const ImageMouseOverEvent =(e)=>{
         e.target.style.cursor = 'pointer';
     }
-    const handleLoginAction=(e)=>{
-        fetch("http://localhost:8080/login")
-    }
     return(
        <div>
         <div className="AppBarBox">
             <img src={logo} alt={logo} className="logoStyle" onClick={()=>{dataFromAppBar("Home")}} onMouseOver={ImageMouseOverEvent} />
             <Movies dataFromMovies={dataFromAppBar}/>
-           <button onClick={handleLoginAction}>login</button>
         </div>
        </div>
     );
