--- Frontend/src/Components/login.js
@@ -1,52 +1,0 @@
-import React, { useState } from 'react';
-
-function Login() {
-  const [username, setUsername] = useState('');
-  const [password, setPassword] = useState('');
-  const [loggedIn, setLoggedIn] = useState(false);
-
-  const handleUsernameChange = (event) => {
-    setUsername(event.target.value);
-  };
-
-  const handlePasswordChange = (event) => {
-    setPassword(event.target.value);
-  };
-
-  const handleLogin = () => {
-    // Simulating login functionality with a hardcoded username and password
-    const validUsername = 'exampleuser';
-    const validPassword = 'examplepassword';
-
-    if (username === validUsername && password === validPassword) {
-      setLoggedIn(true);
-    } else {
-      alert('Invalid credentials. Please try again.');
-    }
-  };
-
-  const handleLogout = () => {
-    setLoggedIn(false);
-    setUsername('');
-    setPassword('');
-  };
-
-  if (loggedIn) {
-    return (
-      <div>
-        <h1>Welcome, {username}!</h1>
-        <button onClick={handleLogout}>Logout</button>
-      </div>
-    );
-  } else {
-    return (
-      <div>
-        <input type="text" placeholder="Username" value={username} onChange={handleUsernameChange} />
-        <input type="password" placeholder="Password" value={password} onChange={handlePasswordChange} />
-        <button onClick={handleLogin}>Login</button>
-      </div>
-    );
-  }
-}
-
-export default Login;
